frameworks/python_server/requirements.txt
gunicorn==19.10.0
gevent==21.1.2
geventhttpclient==1.4.5
greenlet==0.4.17
meinheld==1.0.2
loguru==0.5.3
flask==1.0.2
requests==2.22.0
beautifulsoup4==4.8.0
lxml==4.4.1
joblib==0.13.2
json5==0.8.5
numpy==1.17.2
scipy==1.3.1
contextlib2==0.6.0
PyYAML==5.4.1
packaging==19.0
tqdm==4.64.0
aiohttp==3.8.3
fuzzywuzzy
python-Levenshtein